requirements.txt
# Dépendances pour le projet C8 - Service IA d'Analyse de Sentiment
# Compétence C8 : Paramétrage d'un service d'IA

# Pipeline et modèles Hugging Face
transformers
huggingface_hub
torch

# Monitoring et métriques
prometheus_client

# Configuration et environnement
python-dotenv

# Utilitaires pour affichage et logging
tabulate
tiktoken

# Dépendances optionnelles pour de meilleures performances
# torch-audio  # Pour les modèles audio (optionnel)
# accelerate   # Pour l'accélération GPU (optionnel)

# Dépendances de développement (optionnel)
# pytest        # Pour les tests
# black         # Pour le formatage de code
# flake8        # Pour le linting 